requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.5
django-upload-validator==1.1.5
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
mysqlclient==1.4.6
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-magic==0.4.15
python-magic-bin==0.4.14
pytz==2020.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
websockets==8.1
zope.interface==5.1.0
